final cleanUp 03

+++ src/MoveableShape.java
@@ -1,9 +1,6 @@
-import java.awt.*;
-import java.util.ArrayList;
 
 public interface MoveableShape extends ColoredShape {
 
-    public void move(float dx, float dy);
-
+    public void move(int dx, int dy);
 
 }
